ports/sh: fix long-standing console line counting bug
The computation of how many screen lines at given console lines takes
used the globally-configured font instead of the widget's font. This
cause certain lines to have their size over-estimated on fx-9860G III,
leading to weird scrolling and too few lines being displayed in some
situations.

The bug was present on the fx-CG 50 but did not occur because the font
used for rendering the console is the default font.