numworks: don't set global dwindow in kandinsky
This interferes with the shell and gives fairly bad results with print()
and input(). This way of doing things lets the shell print correctly.
The shell will override the borders, however. It's not immediately
obvious how to fix that properly with the auto-update system in main.c.
I'll try to move the auto screen update back to modkandinsky.c.