ports/sh: enable binary mode on files
Old fdfile code relies on a configuration macro which appears to have
been renamed. The old name being undefined defaulted to a value of 0,
disabling support.

As a result, opening files in binary mode would have no effect, keeping
text mode. This makes no difference for fxlibc (as there's no newline
translation there, as allowed by standard) but this does make a
difference for MicroPython, which then tries to decode all of the input
data as UTF-8.